refactor: db 커넥션 유틸 클래스 위치 이동
`DBConnectionUtils` 클래스의 위치를 database 패키지 하위로 변경함.
